Segment/requirements.txt
-i https://pypi.org/simple
accelerate==1.1.1; python_full_version >= '3.9.0'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.7.0; python_version >= '3.9'
boto3==1.35.76; python_version >= '3.8'
botocore==1.35.76; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.2; python_version >= '3.7'
fastapi==0.115.6; python_version >= '3.8'
filelock==3.16.1; python_version >= '3.8'
fsspec==2024.10.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
huggingface-hub==0.26.2; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
mpmath==1.3.0
networkx==3.2.1; python_version >= '3.9'
numpy==2.0.2; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pillow==11.0.0; python_version >= '3.9'
psutil==6.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.10.3; python_version >= '3.8'
pydantic-core==2.27.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
s3transfer==0.10.4; python_version >= '3.8'
safetensors==0.4.5; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.41.3; python_version >= '3.8'
sympy==1.13.1; python_version >= '3.8'
tokenizers==0.20.3; python_version >= '3.7'
torch==2.5.1; python_full_version >= '3.8.0'
tqdm==4.67.0; python_version >= '3.7'
transformers==4.46.3; python_full_version >= '3.8.0'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==1.26.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.32.1; python_version >= '3.8'
